requirements.txt
autogen-agentchat
autogen-core
autogen-ext[ollama]
python-dotenv
